+++ commands/gsd/debug.md
@@ -4,6 +4,7 @@ description: Systematic debugging with persistent state across context resets
 argument-hint: [list | status <slug> | continue <slug> | --diagnose] [issue description]
 allowed-tools:
   - Read
+  - Write
   - Bash
   - Task
   - AskUserQuestion
