+++ docs/INVENTORY-MANIFEST.json
@@ -304,94 +304,5 @@
       "gsd-validate-commit.sh",
       "gsd-workflow-guard.js"
     ]
-  },
-  "workflows": [
-    "add-backlog.md",
-    "add-phase.md",
-    "add-tests.md",
-    "add-todo.md",
-    "ai-integration-phase.md",
-    "analyze-dependencies.md",
-    "audit-fix.md",
-    "audit-milestone.md",
-    "audit-uat.md",
-    "autonomous.md",
-    "check-todos.md",
-    "cleanup.md",
-    "code-review-fix.md",
-    "code-review.md",
-    "complete-milestone.md",
-    "debug.md",
-    "diagnose-issues.md",
-    "discovery-phase.md",
-    "discuss-phase-assumptions.md",
-    "discuss-phase-power.md",
-    "discuss-phase.md",
-    "do.md",
-    "docs-update.md",
-    "edit-phase.md",
-    "eval-review.md",
-    "execute-phase.md",
-    "execute-plan.md",
-    "explore.md",
-    "extract-learnings.md",
-    "fast.md",
-    "forensics.md",
-    "graduation.md",
-    "health.md",
-    "help.md",
-    "import.md",
-    "inbox.md",
-    "ingest-docs.md",
-    "insert-phase.md",
-    "list-phase-assumptions.md",
-    "list-workspaces.md",
-    "manager.md",
-    "map-codebase.md",
-    "milestone-summary.md",
-    "new-milestone.md",
-    "new-project.md",
-    "new-workspace.md",
-    "next.md",
-    "node-repair.md",
-    "note.md",
-    "pause-work.md",
-    "plan-milestone-gaps.md",
-    "plan-phase.md",
-    "plan-review-convergence.md",
-    "plant-seed.md",
-    "pr-branch.md",
-    "profile-user.md",
-    "progress.md",
-    "quick.md",
-    "reapply-patches.md",
-    "remove-phase.md",
-    "remove-workspace.md",
-    "resume-project.md",
-    "review.md",
-    "scan.md",
-    "secure-phase.md",
-    "session-report.md",
-    "settings-advanced.md",
-    "settings-integrations.md",
-    "settings.md",
-    "ship.md",
-    "sketch-wrap-up.md",
-    "sketch.md",
-    "spec-phase.md",
-    "spike-wrap-up.md",
-    "spike.md",
-    "stats.md",
-    "sync-skills.md",
-    "thread.md",
-    "transition.md",
-    "ui-phase.md",
-    "ui-review.md",
-    "ultraplan-phase.md",
-    "undo.md",
-    "update.md",
-    "validate-phase.md",
-    "verify-phase.md",
-    "verify-work.md"
-  ]
+  }
 }
